+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added Payloads creator into API class
 - Added $currentUsername parameter in the "updateUser" method into UserEndpoint
 - Fixed a bug of the "createSystemProperty" and "updateSystemProperty" methods into SystemEndpoint
+- Added $servicename parameter to ChatRooms Endpoint methods
 
 ## 1.2.2 
 - Added Basic Authentication